-Added function for calculating cable cross-section -Added function for calculating Switch power -Added function for creating connections with wires -Added class Connection, Cable, Position, Connector -Json and console inputs are not edited!

+++ FuseBox/App/Models/BaseElectrical.cs
@@ -6,6 +6,6 @@ namespace FuseBox.App.Models
     public abstract class BaseElectrical : BaseEntity
     {
         [JsonProperty(Order = 3)]
-        public int Amper { get; set; }
+        public double Amper { get; set; }
     }
 }
